fix circular dependencies in build
