+++ packages/back-end/test/services/features-holdout.test.ts
@@ -31,15 +31,10 @@ import {
   updateSDKPayload,
 } from "back-end/src/models/SdkPayloadModel";
 
-// Mock shared/util - use same pattern as features-public to avoid circular dep
-jest.mock("shared/util", () => {
-  const actual =
-    jest.requireActual<typeof import("shared/util")>("shared/util");
-  return {
-    ...actual,
-    getSavedGroupsValuesFromInterfaces: jest.fn().mockReturnValue({}),
-  };
-});
+jest.mock("shared/util", () => ({
+  ...jest.requireActual<typeof import("shared/util")>("shared/util"),
+  getSavedGroupsValuesFromInterfaces: jest.fn().mockReturnValue({}),
+}));
 
 // Mock config to prevent MongoDB log messages
 jest.mock("back-end/src/init/config", () => ({
